refactor: adopt jest projects structure
